# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MLSystem2 Training UI API",
    "description": "Public FastAPI contract for the MLSystem2 training site.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/training-jobs": {
      "post": {
        "summary": "Post Training Job",
        "operationId": "post_training_job_api_v1_training_jobs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrainingJobCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/queues/training/enabled": {
      "put": {
        "summary": "Put Training Enabled",
        "operationId": "put_training_enabled_api_v1_queues_training_enabled_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueueEnabledUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueueSnapshot"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/queues/inference/enabled": {
      "put": {
        "summary": "Put Inference Enabled",
        "operationId": "put_inference_enabled_api_v1_queues_inference_enabled_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueueEnabledUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueueSnapshot"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/jobs/{job_id}": {
      "get": {
        "summary": "Get Job",
        "operationId": "get_job_api_v1_jobs__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Job Route",
        "operationId": "delete_job_route_api_v1_jobs__job_id__delete",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/jobs/{job_id}/log": {
      "get": {
        "summary": "Get Job Log",
        "operationId": "get_job_log_api_v1_jobs__job_id__log_get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobLogInfo"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/jobs/{job_id}/move-up": {
      "post": {
        "summary": "Move Job Up",
        "operationId": "move_job_up_api_v1_jobs__job_id__move_up_post",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/jobs/{job_id}/move-down": {
      "post": {
        "summary": "Move Job Down",
        "operationId": "move_job_down_api_v1_jobs__job_id__move_down_post",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/results/classes/{class_key}": {
      "get": {
        "summary": "Get Class Results",
        "operationId": "get_class_results_api_v1_results_classes__class_key__get",
        "parameters": [
          {
            "name": "class_key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Class Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClassResultsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/results/classes/{class_key}/pseudo-markup": {
      "post": {
        "summary": "Post Pseudo Markup",
        "operationId": "post_pseudo_markup_api_v1_results_classes__class_key__pseudo_markup_post",
        "parameters": [
          {
            "name": "class_key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Class Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_post_pseudo_markup_api_v1_results_classes__class_key__pseudo_markup_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/results/pseudo-markup/{result_id}": {
      "delete": {
        "summary": "Delete Pseudo Markup",
        "operationId": "delete_pseudo_markup_api_v1_results_pseudo_markup__result_id__delete",
        "parameters": [
          {
            "name": "result_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Result Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PseudoMarkupResultInfo"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_post_pseudo_markup_api_v1_results_classes__class_key__pseudo_markup_post": {
        "properties": {
          "dataset_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dataset Key"
          },
          "image_folder_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Folder Key"
          },
          "training_result_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Training Result Id"
          },
          "scenes_txt": {
            "anyOf": [
              {
                "type": "string",
                "format": "binary"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scenes Txt"
          }
        },
        "type": "object",
        "title": "Body_post_pseudo_markup_api_v1_results_classes__class_key__pseudo_markup_post"
      },
...
      "ClassResultsResponse": {
        "properties": {
          "class_key": {
            "type": "string",
            "title": "Class Key"
          },
          "class_name": {
            "type": "string",
            "title": "Class Name"
          },
          "dataset_updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dataset Updated At"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/TrainingResultInfo"
            },
            "type": "array",
            "title": "Results"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "class_key",
          "class_name",
          "results"
        ],
        "title": "ClassResultsResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JobDetail": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "type": {
            "$ref": "#/components/schemas/JobType"
          },
          "source": {
            "$ref": "#/components/schemas/JobSource",
            "default": "manual"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          },
          "queue_position": {
            "type": "integer",
            "title": "Queue Position"
          },
          "dataset_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dataset Key"
          },
          "dataset_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dataset Version"
          },
          "dataset_name": {
            "type": "string",
            "title": "Dataset Name"
          },
          "training_dataset_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Training Dataset Name"
          },
          "inference_dataset_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Inference Dataset Name"
          },
          "model_name": {
            "type": "string",
            "title": "Model Name"
          },
          "architecture": {
            "type": "string",
            "title": "Architecture"
          },
          "tile_size": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tile Size"
          },
          "mlflow_experiment_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mlflow Experiment Name"
          },
          "mlflow_run_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mlflow Run Name"
          },
          "config": {
            "additionalProperties": true,
            "type": "object",
            "title": "Config"
          },
          "readonly": {
            "type": "boolean",
            "title": "Readonly",
            "default": true
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "progress": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RuntimeProgress"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "type",
          "status",
          "queue_position",
          "dataset_name",
          "model_name",
          "architecture",
          "config",
          "created_at"
        ],
        "title": "JobDetail"
      },
      "JobLogInfo": {
        "properties": {
          "job_id": {
            "type": "string",
            "format": "uuid",
            "title": "Job Id"
          },
          "source_name": {
            "type": "string",
            "title": "Source Name"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "truncated": {
            "type": "boolean",
            "title": "Truncated",
            "default": false
          },
          "size_bytes": {
            "type": "integer",
            "title": "Size Bytes"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "job_id",
          "source_name",
          "content",
          "size_bytes"
        ],
        "title": "JobLogInfo"
      },
      "JobSource": {
        "type": "string",
        "enum": [
          "manual",
          "automation"
        ],
        "title": "JobSource"
      },
      "JobStatus": {
        "type": "string",
        "enum": [
          "queued",
          "running",
          "completed",
          "failed",
          "cancelled"
        ],
        "title": "JobStatus"
      },
...
      "JobType": {
        "type": "string",
        "enum": [
          "training",
          "inference"
        ],
        "title": "JobType"
      },
...
      "PseudoMarkupResultInfo": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "job_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Id"
          },
          "source": {
            "$ref": "#/components/schemas/JobSource",
            "default": "manual"
          },
          "dataset_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dataset Key"
          },
          "dataset_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dataset Version"
          },
          "source_dataset_name": {
            "type": "string",
            "title": "Source Dataset Name"
          },
          "scenes_file": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/StoredFileInfo"
              },
              {
                "type": "null"
              }
            ]
          },
          "geojson_file": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/StoredFileInfo"
              },
              {
                "type": "null"
              }
            ]
          },
          "image_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Count"
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "running",
              "ok",
              "error",
              "cancelled"
            ],
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "progress": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RuntimeProgress"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "source_dataset_name",
          "status",
          "created_at"
        ],
        "title": "PseudoMarkupResultInfo"
      },
      "QueueEnabledUpdate": {
        "properties": {
          "enabled": {
            "type": "boolean",
            "title": "Enabled"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "enabled"
        ],
        "title": "QueueEnabledUpdate"
      },
      "QueueSnapshot": {
        "properties": {
          "training_enabled": {
            "type": "boolean",
            "title": "Training Enabled"
          },
          "inference_enabled": {
            "type": "boolean",
            "title": "Inference Enabled"
          },
          "jobs": {
            "items": {
              "$ref": "#/components/schemas/JobSummary"
            },
            "type": "array",
            "title": "Jobs"
          },
          "training_jobs": {
            "items": {
              "$ref": "#/components/schemas/JobSummary"
            },
            "type": "array",
            "title": "Training Jobs"
          },
          "inference_jobs": {
            "items": {
              "$ref": "#/components/schemas/JobSummary"
            },
            "type": "array",
            "title": "Inference Jobs"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "training_enabled",
          "inference_enabled",
          "training_jobs",
          "inference_jobs"
        ],
        "title": "QueueSnapshot"
      },
...
      "RuntimeProgress": {
        "properties": {
          "current": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current"
          },
          "total": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total"
          },
          "elapsed_minutes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Elapsed Minutes"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "RuntimeProgress"
      },
      "StoredFileInfo": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "kind": {
            "$ref": "#/components/schemas/StoredFileKind"
          },
          "original_name": {
            "type": "string",
            "title": "Original Name"
          },
          "size_bytes": {
            "type": "integer",
            "title": "Size Bytes"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "download_url": {
            "type": "string",
            "title": "Download Url"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "kind",
          "original_name",
          "size_bytes",
          "created_at",
          "download_url"
        ],
        "title": "StoredFileInfo"
      },
      "StoredFileKind": {
        "type": "string",
        "enum": [
          "scenes_txt",
          "annotation_geojson",
          "pseudo_markup_geojson"
        ],
        "title": "StoredFileKind"
      },
...
      "TrainingJobCreate": {
        "properties": {
          "mlflow_experiment_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mlflow Experiment Id"
          },
          "mlflow_experiment_name": {
            "type": "string",
            "title": "Mlflow Experiment Name"
          },
          "mlflow_run_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mlflow Run Name"
          },
          "dataset_key": {
            "type": "string",
            "title": "Dataset Key"
          },
          "custom_dataset_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Custom Dataset Id"
          },
          "architecture": {
            "type": "string",
            "title": "Architecture"
          },
          "config": {
            "additionalProperties": true,
            "type": "object",
            "title": "Config"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "mlflow_experiment_name",
          "dataset_key",
          "architecture",
          "config"
        ],
        "title": "TrainingJobCreate"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}